+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 1.14 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added function to construct link to object 
+  [sgeulette]
 
 1.13 (2017-11-27)
 -----------------