+++ CHANGES.rst
@@ -5,9 +5,11 @@ Changelog
 1.14 (unreleased)
 -----------------
 
-- Added function to construct link to object 
+- Added function to construct link to object
   [sgeulette]
 - Added function to set an attribute on context
+  [sgeulette]
+- Improved dv_conversion
   [sgeulette]
 
 1.13 (2017-11-27)